+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - guides were a code smell, so they've been removed.  Instead we have EnumValue
 which allows you to specify a unit enum with an additional value argument that
 is stored in the schema JSON
+- EnumValue now returns an Option type to handle Unspecified
 
 ## [0.1.4]
 - fixed a bug in api::call that was affecting calls with more than one argument